SRE-3945: Add service impacts to Datadog alerts in Slack
This commit enriches the Datadog alert messages forwarded to Slack. Alert titles are now bolded, the date is included, as well as the alert priority and scope. There's also now a link to the specific event, and the affected services are listed. This change allows for quicker and more efficient responses to alerts. In the Datadog processor, the alert structure now includes 'Services', indicating which services are affected by the alert. The processor also now queries a Prometheus endpoint to get a list of services affected by the product specified in the alert scope.

In the Slack message template, it now extracts the channel and thread IDs from the same GitLab pipeline variables as before, but under updated identifiers. Moreover, it tries to fetch chart images by checking more snapshot URLs than before. These enrichments provide much-needed context right in the Slack message, removing the need to jump back to Datadog or GitLab for additional investigation.